# docs/openapi/creators-api.yaml
...
openapi: 3.1.0
info:
  title: qoe.fi — API Créateurs & Médias
  version: 1.0.0
  description: >-
    API pour les médias et créateurs qui utilisent qoe.fi comme plateforme de
    diffusion (publier sur qoe.fi depuis leur CMS) ou comme CMS headless
    (diffuser leurs articles sur leur propre site). Contrat stable — les
    changements sont versionnés, jamais cassants.
  contact:
    name: qoe.fi
    url: https://qoe.fi
servers:
  - url: https://api.qoe.fi
    description: Production (backend Go)
  - url: http://localhost:8080
    description: Développement local
...
security:
  - apiKeyAuth: []
paths:
...
  /v1/categories:
    get:
      tags: [categories]
      operationId: listCategories
      summary: Liste les catégories de la publication avec le comptage d'articles publiés
      responses:
        '200':
          description: Catégories
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/analytics/stats:
    get:
      tags: [analytics]
      operationId: getAnalyticsStats
      summary: Statistiques de lecture (proxy Umami)
      parameters:
        - name: startAt
          in: query
          description: 'Timestamp ms de début (défaut : 30 jours)'
          schema:
            type: integer
            format: int64
        - name: endAt
          in: query
          description: 'Timestamp ms de fin (défaut : maintenant)'
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Stats + pages les plus lues
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [stats, topPages]
                    properties:
                      stats:
                        type: object
                        additionalProperties: true
                      topPages:
                        type: array
                        items:
                          type: object
                          additionalProperties: true
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    apiKeyAuth:
      type: http
      scheme: bearer
      description: >-
        Clé API `qoe_live_*` (délivrée après revue manuelle du média). Scopes
        (moindre privilège) : READ (articles/catégories), WRITE
        (créer/modifier/publier/supprimer), ANALYTICS (stats). Une clé n'a
        accès qu'aux endpoints de ses scopes — 403 sinon.
    jwtAuth:
      type: http
      scheme: bearer
      description: JWT Supabase (app mobile)
    oauth2: # ⚠️ à construire
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.qoe.fi/authorize
          tokenUrl: https://auth.qoe.fi/token
          scopes:
            articles:read: Lire les articles
            articles:write: Créer/modifier des articles
            analytics:read: Lire les statistiques
...
  responses:
...
    Unauthorized:
      description: Clé API / JWT manquant ou invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Category:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: [string, 'null']
        articlesCount:
          type: integer
          description: Nombre d'articles publiés dans la catégorie
...